// src/Interfaces/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  graphql_public: {
    Tables: {
      [_ in never]: never
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      graphql: {
        Args: {
          operationName?: string
          query?: string
          variables?: Json
          extensions?: Json
        }
        Returns: Json
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  public: {
    Tables: {
      Articles: {
        Row: {
          article_id: number
          EAN: number | null
          Name: string | null
          PCS: number | null
          Size: string | null
          Unit: string | null
        }
        Insert: {
          article_id: number
          EAN?: number | null
          Name?: string | null
          PCS?: number | null
          Size?: string | null
          Unit?: string | null
        }
        Update: {
          article_id?: number
          EAN?: number | null
          Name?: string | null
          PCS?: number | null
          Size?: string | null
          Unit?: string | null
        }
        Relationships: []
      }
      inventories: {
        Row: {
          article_id: number | null
          inventory_id: string
          last_updated: string | null
          location_id: string | null
          quantity: number | null
        }
        Insert: {
          article_id?: number | null
          inventory_id?: string
          last_updated?: string | null
          location_id?: string | null
          quantity?: number | null
        }
        Update: {
          article_id?: number | null
          inventory_id?: string
          last_updated?: string | null
          location_id?: string | null
          quantity?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "inventories_article_id_fkey"
            columns: ["article_id"]
            isOneToOne: false
            referencedRelation: "Articles"
            referencedColumns: ["article_id"]
          },
          {
            foreignKeyName: "inventories_location_id_fkey"
            columns: ["location_id"]
            isOneToOne: false
            referencedRelation: "Locations"
            referencedColumns: ["location_id"]
          },
        ]
      }
      Locations: {
        Row: {
          aisle: string | null
          bin: string | null
          location_code: string | null
          location_id: string
          location_type: string | null
          max_capacity: number | null
          rack: string | null
          shelf: string | null
          warehouse: string | null
          zone: string | null
        }
        Insert: {
          aisle?: string | null
          bin?: string | null
          location_code?: string | null
          location_id?: string
          location_type?: string | null
          max_capacity?: number | null
          rack?: string | null
          shelf?: string | null
          warehouse?: string | null
          zone?: string | null
        }
        Update: {
          aisle?: string | null
          bin?: string | null
          location_code?: string | null
          location_id?: string
          location_type?: string | null
          max_capacity?: number | null
          rack?: string | null
          shelf?: string | null
          warehouse?: string | null
          zone?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never